extract openAL specifics into separate file
